Extract attribute list component
