+++ src/main/scripts/src/app/collection/config/attribute-list/configured-attribute.ts
