+++ src/main/scripts/src/app/collection/config/attribute-list/constraint-suggestion.ts
