+++ src/main/scripts/src/app/collection/config/attribute-list/constraints.ts
